+++ src/pages/home/information/Information.tsx
@@ -6,14 +6,6 @@ import { useTranslation } from 'react-i18next';
 
 type Category = '주의사항' | '안전교육' | '의료기관' | '상담' | '신고';
 
-const categoryMap = {
-  주의사항: 'category.caution',
-  안전교육: 'category.safety',
-  의료기관: 'category.hospital',
-  상담: 'category.consult',
-  신고: 'category.report',
-} as const;
-
 const Information = () => {
   const [isClicked, setIsClicked] = useState<Category>('주의사항');
   const { t } = useTranslation();
